Enforce minimum width for numberline width
Modified the random width generation code to take in a correct minimum
width of the number line.

The minimum width is the end_unit * width_low * width_mod. This way if
the start_unit is 1 and the end_unit is 10, the 1 and 10 will be a
minimum of width_mod * 20px (based on current width_low) in size.