Calculate handle shift offsets
Calculate the number of pixels that a handle may have to be offset in
order for number line to be perceptionally intact.

Offsets are required because the guides have a width. This will mess up
how a subject perceives the line. The offset adds a pixel so the handle
may be placed in the correct spot and look correct.