Set configuration file extension to `nlconfig`
The configuration file did not have an extension prior to this commit.
It was a normal text file. This commit sets it to `nlconfig`.

A non-standard extension was picked for two reasons:

 1. Discourage direct editing of file through soft barrier of access
 2. Make the configuration instantly recognisable.

To accommodate this change, both the save and load dialogs have filters
to exclude all other file formats.