Refactor target UI design and error checking
Refactored the user interface for the target panel to contain the
distinct targets and whether or not to include the left or right bounds
in the pool of target values.

In addition, error checking has been make continuous, which allows the
experimenter to see the errors before hitting okay on the detailed
configuration panel.